requirements.txt
setuptools==57.0.0
wheel==0.36.2
django==4.1.3
djangorestframework==3.14.0
pillow==9.3.0
djangorestframework-simplejwt==5.2.2
datetime==4.7
python-dateutil==2.8.2
geopy==2.2.0
certifi==2022.9.24
pause==0.3
django-cors-headers==3.13.0
django-base64field==1.0
